014 Moved work into model directory
